move useToday into its own file
